Add images to help pages
We are using some of the images from gnome-user-docs. I don't think
its worth the complexity and effort to copy every asset used in both
to this repo as well. Especially as they both install to the same
spot (/usr/share/help/gnome-help).

But user docs will need to be installed alongside this for all the
images to work
[endlessm/eos-shell#4235]