Add translation support for a few new languages
Includes integration with Transifex.

Unlike other packages that deliver translations as .mo files
that are essentially empty if no translations, this package
delivers translations as a full set of .page files that contain
all the original English strings if no translations.  For this
reason, we are not including the full superset of languages,
but only adding those that are supported upstream or that we
exlicitly want to support at this time.

https://phabricator.endlessm.com/T13620